rename sentryUtils to sentry/utils
